requirements.txt
fastapi
uvicorn
pillow
python-multipart
jinja2
